fix: float 32 literals, feat: structs no longer need type provided/annotation

+++ include/pico/binding/address_env.h
@@ -36,7 +36,7 @@ typedef struct {
     AddressEntry_t type;
     union {
         void* value;
-        int8_t stack_offset;
+        int16_t stack_offset;
     };
 } AddressEntry;
